📮 Maintain preamble as structured data in myst-to-tex with funcitons to reconstruct later
