+++ .changeset/cold-pans-lay.md
@@ -1,0 +1,5 @@
+---
+'myst-to-tex': patch
+---
+
+Maintain preamble as structured data in myst-to-tex with funcitons to reconstruct later
