+++ packages/myst-to-tex/tests/glossaries.spec.ts
@@ -4,7 +4,7 @@ import path from 'node:path';
 import yaml from 'js-yaml';
 import { unified } from 'unified';
 import type { LatexResult } from '../src';
-import mystToTex from '../src';
+import mystToTex, { generatePreamble } from '../src';
 
 type TestCase = {
   title: string;
@@ -31,13 +31,18 @@ const casesList: TestCases[] = fs
 
 casesList.forEach(({ title, latexGlossary, cases }) => {
   describe(title, () => {
-    test.each(cases.map((c): [string, string, TestCase] => [c.title, latexGlossary, c]))('%s', (_, expectedLatexGlossary, { latex, mdast, printGlossaries }) => {
-      const pipe = unified().use(mystToTex, { printGlossaries });
-      pipe.runSync(mdast as any);
-      const file = pipe.stringify(mdast as any);
-      const printedGlossary = printGlossaries ? `\n${expectedLatexGlossary}` : '';
-      const expectedResult = `${latex}${printedGlossary}`;
-      expect((file.result as LatexResult).value).toEqual(expectedResult);
-    });
+    test.each(cases.map((c): [string, string, TestCase] => [c.title, latexGlossary, c]))(
+      '%s',
+      (_, expectedLatexGlossary, { latex, mdast, printGlossaries }) => {
+        const pipe = unified().use(mystToTex, { printGlossaries });
+        pipe.runSync(mdast as any);
+        const file = pipe.stringify(mdast as any);
+        const { suffix } = generatePreamble((file.result as LatexResult).preamble);
+        const latexValue = (file.result as LatexResult).value + suffix;
+        const printedGlossary = printGlossaries ? `\n${expectedLatexGlossary}` : '';
+        const expectedResult = `${latex}${printedGlossary}`;
+        expect(latexValue).toEqual(expectedResult);
+      },
+    );
   });
 });
